+++ administrator/components/com_search/Controller/Controller.php
@@ -6,15 +6,19 @@
  * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
+namespace Joomla\Component\Search\Administrator\Controller;
 
 defined('_JEXEC') or die;
+
+use Joomla\CMS\Controller\Controller as BaseController;
+use Joomla\Component\Search\Administrator\Helper\SearchHelper;
 
 /**
  * Search master display controller.
  *
  * @since  1.6
  */
-class SearchController extends JControllerLegacy
+class Controller extends BaseController
 {
 	/**
 	 * @var		string	The default view.
@@ -26,15 +30,15 @@ class SearchController extends JControllerLegacy
 	 * Method to display a view.
 	 *
 	 * @param   boolean  $cachable   If true, the view output will be cached
-	 * @param   array    $urlparams  An array of safe URL parameters and their variable types, for valid values see {@link JFilterInput::clean()}.
+	 * @param   array    $urlparams  An array of safe URL parameters and their variable types, for valid values see {@link \JFilterInput::clean()}.
 	 *
-	 * @return  SearchController  This object to support chaining.
+	 * @return  static  This object to support chaining.
 	 *
 	 * @since   1.5
 	 */
 	public function display($cachable = false, $urlparams = false)
 	{
-		JLoader::register('SearchHelper', JPATH_ADMINISTRATOR . '/components/com_search/helpers/search.php');
+		\JLoader::register('SearchHelper', JPATH_ADMINISTRATOR . '/components/com_search/helpers/search.php');
 
 		// Load the submenu.
 		SearchHelper::addSubmenu($this->input->get('view', 'searches'));
