+++ components/com_search/View/Search/Opensearch.php
@@ -6,47 +6,50 @@
  * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
+namespace Joomla\Component\Search\Site\View\Search;
 
 defined('_JEXEC') or die;
+
+use Joomla\CMS\Component\ComponentHelper;
+use Joomla\CMS\View\AbstractView;
 
 /**
  * OpenSearch View class for the Search component
  *
  * @since  1.7
  */
-class SearchViewSearch extends JViewLegacy
+class SearchViewSearch extends AbstractView
 {
 	/**
 	 * Execute and display a template script.
 	 *
 	 * @param   string  $tpl  name of the template
 	 *
-	 * @throws Exception
+	 * @throws \Exception
 	 * @return  mixed  A string if successful, otherwise an Error object.
 	 */
 	public function display($tpl = null)
 	{
-		$doc = JFactory::getDocument();
-		$app = JFactory::getApplication();
+		$doc = \JFactory::getDocument();
+		$app = \JFactory::getApplication();
 
-		$params = JComponentHelper::getParams('com_search');
+		$params = ComponentHelper::getParams('com_search');
 		$doc->setShortName($params->get('opensearch_name', $app->get('sitename')));
 		$doc->setDescription($params->get('opensearch_description', $app->get('MetaDesc')));
 
 		// Add the URL for the search
-		$searchUri = JUri::base() . 'index.php?option=com_search&searchword={searchTerms}';
+		$searchUri = \JUri::base() . 'index.php?option=com_search&searchword={searchTerms}';
 
 		// Find the menu item for the search
-		$menu  = $app->getMenu();
-		$items = $menu->getItems('link', 'index.php?option=com_search&view=search');
+		$items = $app->getMenu()->getItems('link', 'index.php?option=com_search&view=search');
 
 		if (isset($items[0]))
 		{
 			$searchUri .= '&Itemid=' . $items[0]->id;
 		}
 
-		$htmlSearch           = new JOpenSearchUrl;
-		$htmlSearch->template = JRoute::_($searchUri);
+		$htmlSearch           = new \JOpenSearchUrl;
+		$htmlSearch->template = \JRoute::_($searchUri);
 		$doc->addUrl($htmlSearch);
 	}
 }
