+++ components/com_search/dispatcher.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * @package     Joomla.Site
+ * @subpackage  com_search
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Dispatcher\Dispatcher;
+
+/**
+ * Dispatcher class for com_search
+ *
+ * @since  __DEPLOY_VERSION__
+ */
+class SearchDispatcher extends Dispatcher
+{
+	/**
+	 * The extension namespace
+	 *
+	 * @var    string
+	 *
+	 * @since  __DEPLOY_VERSION__
+	 */
+	protected $namespace = 'Joomla\\Component\\Search';
+}
